requirements.txt
streamlit
ultralytics
torch
torchvision
timm
Pillow
numpy
